+++ cmd/substreams/sink-deploy.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"fmt"
-	server "github.com/streamingfast/substreams/sink-server"
+	"github.com/streamingfast/substreams/sink-server/context"
+	"github.com/streamingfast/substreams/sink-server/printer"
 	"net/http"
 
 	"github.com/bufbuild/connect-go"
@@ -49,7 +50,7 @@ func deployE(cmd *cobra.Command, args []string) error {
 
 	fmt.Printf("Deploying... (creating services, please wait)\n")
 	cli := pbsinksvcconnect.NewProviderClient(http.DefaultClient, sflags.MustGetString(cmd, "endpoint"))
-	ctx = server.WithDeployStatusPrinter(ctx, &server.DeployStatusPrinter{})
+	ctx = context.WithDeployStatusPrinter(ctx, &printer.DeployStatusPrinter{})
 
 	resp, err := cli.Deploy(ctx, connect.NewRequest(req))
 	if err != nil {
