+++ sink-server/context/context.go
@@ -1,22 +1,25 @@
-package server
+package context
 
-import "context"
+import (
+	"context"
+	"github.com/streamingfast/substreams/sink-server/printer"
+)
 
 type sinkContextKey int
 
 const deployStatusPrinterKey = sinkContextKey(0)
 
-func WithDeployStatusPrinter(ctx context.Context, printer *DeployStatusPrinter) context.Context {
+func WithDeployStatusPrinter(ctx context.Context, printer *printer.DeployStatusPrinter) context.Context {
 	return context.WithValue(ctx, deployStatusPrinterKey, printer)
 }
 
 func GetDeployStatusPrinter(ctx context.Context) interface {
 	Printf(format string, args ...interface{})
 } {
-	if printer, ok := ctx.Value(deployStatusPrinterKey).(*DeployStatusPrinter); ok {
+	if printer, ok := ctx.Value(deployStatusPrinterKey).(*printer.DeployStatusPrinter); ok {
 		return printer
 	}
-	return defaultDeployStatusPrinter
+	return printer.DefaultDeployStatusPrinter
 }
 
 const productionModeKey = sinkContextKey(1)
